chore(server): finish scoping test from socketOnDisconnection big file into small one
