+++ packages/server/tests/SocketonDisconnection.test.ts
@@ -1,19 +1,13 @@
 import Database from '@ioc:Adonis/Lucid/Database';
-import {
-    MtvWorkflowState,
-    MtvWorkflowStateWithUserRelatedInformation,
-} from '@musicroom/types';
+import { MtvWorkflowStateWithUserRelatedInformation } from '@musicroom/types';
 import ServerToTemporalController from 'App/Controllers/Http/Temporal/ServerToTemporalController';
 import Device from 'App/Models/Device';
 import MtvRoom from 'App/Models/MtvRoom';
-import User from 'App/Models/User';
-import SocketLifecycle from 'App/Services/SocketLifecycle';
-import UserService from 'App/Services/UserService';
 import { datatype, name, random } from 'faker';
 import test from 'japa';
 import sinon from 'sinon';
 import supertest from 'supertest';
-import InitTestUtils, { BASE_URL, sleep } from './TestUtils';
+import InitTestUtils, { BASE_URL, sleep } from './utils/TestUtils';
 /**
  * User should create a room, and removes it after user disconnection
  * User should join a room
@@ -141,59 +135,6 @@ test.group('Rooms life cycle', (group) => {
         await disconnectSocket(socket);
         const roomAfter = await MtvRoom.findBy('creator', userID);
         assert.isNull(roomAfter);
-    });
-
-    test('It should handle a temporal onCreate error', async (assert) => {
-        const userID = datatype.uuid();
-        const socket = await createUserAndGetSocket({ userID });
-        let roomID: undefined | string;
-        const receivedEvents: string[] = [];
-        socket.once('CREATE_ROOM_CALLBACK', () => {
-            receivedEvents.push('CREATE_ROOM_CALLBACK');
-        });
-        const roomName = random.words(1);
-
-        /** Mocks */
-        sinon
-            .stub(ServerToTemporalController, 'createMtvWorkflow')
-            .callsFake(async ({ workflowID }) => {
-                roomID = workflowID;
-
-                /**
-                 * Checking if the user is well registered in the socket-io
-                 * room instance
-                 */
-                const connectedSockets =
-                    await SocketLifecycle.getConnectedSocketToRoom(workflowID);
-                assert.isTrue(connectedSockets.has(socket.id));
-
-                throw new Error('Mocked error');
-            });
-        /** ***** */
-
-        /**
-         * Emit CREATE_ROOM
-         */
-        socket.emit('CREATE_ROOM', { name: roomName, initialTracksIDs: [] });
-        await sleep();
-
-        if (roomID === undefined) throw new Error('roomID is undefined');
-        /**
-         * Checking if the user has correctly been removed from the
-         * room instance
-         */
-        const connectedSockets = await SocketLifecycle.getConnectedSocketToRoom(
-            roomID,
-        );
-        assert.isFalse(connectedSockets.has(socket.id));
-        assert.equal(connectedSockets.size, 0);
-
-        /**
-         * Even if the room hasn't been inserted checking
-         * if it's really not in postgres
-         */
-        const roomBefore = await MtvRoom.findBy('creator', userID);
-        assert.isNull(roomBefore);
     });
 
     test('When a room got evicted, users in it should receive a FORCED_DISCONNECTION socket event', async (assert) => {
@@ -363,409 +304,5 @@ test.group('Rooms life cycle', (group) => {
          * Check if room is not in database
          */
         assert.isNull(await MtvRoom.findBy('creator', userID));
-    });
-    test(`It should creates a room, and it should join room for every user's session/device after one emits JOIN_ROOM
-    It should send USER_LENGTH_UPDATE to every already existing room members`, async (assert) => {
-        const userID = datatype.uuid();
-        const creatorID = datatype.uuid();
-        const roomName = random.word();
-        let userCouldEmitAnExclusiveRoomSignal = false;
-        let state: MtvWorkflowStateWithUserRelatedInformation | undefined;
-
-        /** Mocks */
-        sinon
-            .stub(ServerToTemporalController, 'createMtvWorkflow')
-            .callsFake(async ({ workflowID }) => {
-                state = {
-                    currentTrack: null,
-                    roomID: workflowID,
-                    roomCreatorUserID: datatype.uuid(),
-                    tracks: [
-                        {
-                            id: datatype.uuid(),
-                            artistName: name.findName(),
-                            duration: 42000,
-                            title: random.words(3),
-                            score: datatype.number(),
-                        },
-                    ],
-                    suggestedTracks: null,
-                    playing: false,
-                    name: roomName,
-                    userRelatedInformation: {
-                        userID: creatorID,
-                        emittingDeviceID: datatype.uuid(),
-                    },
-                    usersLength: 1,
-                };
-
-                return {
-                    runID: datatype.uuid(),
-                    workflowID: workflowID,
-                    state,
-                };
-            });
-
-        sinon
-            .stub(ServerToTemporalController, 'joinWorkflow')
-            .callsFake(async () => {
-                if (state === undefined) throw new Error('State is undefined');
-                state.usersLength++;
-                await supertest(BASE_URL)
-                    .post('/temporal/user-length-update')
-                    .send({
-                        ...state,
-                        userRelatedInformation: null,
-                    });
-                await supertest(BASE_URL)
-                    .post('/temporal/join')
-                    .send({ state, joiningUserID: userID });
-                return;
-            });
-
-        sinon.stub(ServerToTemporalController, 'play').callsFake(async () => {
-            userCouldEmitAnExclusiveRoomSignal = true;
-            return;
-        });
-        /** ***** */
-
-        /**
-         * Fisrt creatorUser creates a room
-         */
-        const creatorUser = await createUserAndGetSocket({
-            userID: creatorID,
-        });
-        const creatorReceivedEvents: string[] = [];
-        creatorUser.emit('CREATE_ROOM', {
-            name: random.word(),
-            initialTracksIDs: [datatype.uuid()],
-        });
-        await sleep();
-        const createdRoom = await MtvRoom.findBy('creator', creatorID);
-        assert.isNotNull(createdRoom);
-        if (!createdRoom) throw new Error('room is undefined');
-
-        /**
-         * JoiningUser connects 2 socket and joins the createdRoom with one
-         */
-        creatorUser.once('USER_LENGTH_UPDATE', () => {
-            creatorReceivedEvents.push('USER_LENGTH_UPDATE');
-        });
-
-        const receivedEvents: string[] = [];
-        const joiningUser = {
-            socketA: await createUserAndGetSocket({ userID }),
-            socketB: await createSocketConnection({ userID }),
-        };
-        Object.values(joiningUser).forEach((socket, i) =>
-            socket.once('JOIN_ROOM_CALLBACK', () => {
-                receivedEvents.push(`JOIN_ROOM_CALLBACK_${i}`);
-            }),
-        );
-        joiningUser.socketA.emit('JOIN_ROOM', { roomID: createdRoom.uuid });
-        await sleep();
-        console.log(receivedEvents);
-        assert.isTrue(receivedEvents.includes('JOIN_ROOM_CALLBACK_0'));
-        assert.isTrue(receivedEvents.includes('JOIN_ROOM_CALLBACK_1'));
-        assert.equal(creatorReceivedEvents.length, 1);
-
-        /**
-         * JoiningUser emit an ACTION_PLAY with the other socket connection
-         * It achieves only if he joined the socket io server room
-         */
-        joiningUser.socketB.emit('ACTION_PLAY');
-        await sleep();
-        assert.equal(userCouldEmitAnExclusiveRoomSignal, true);
-    });
-    test("It should join room for every user's session/device after one emits CREATE_ROOM", async (assert) => {
-        const roomName = random.word();
-        const userID = datatype.uuid();
-        let userCouldEmitAnExclusiveRoomSignal = false;
-
-        /** Mocks */
-        sinon
-            .stub(ServerToTemporalController, 'createMtvWorkflow')
-            .callsFake(async ({ workflowID }) => {
-                return {
-                    runID: datatype.uuid(),
-                    workflowID,
-                    state: {
-                        roomID: workflowID,
-                        roomCreatorUserID: datatype.uuid(),
-                        playing: false,
-                        name: roomName,
-                        currentTrack: null,
-                        userRelatedInformation: {
-                            userID,
-                            emittingDeviceID: datatype.uuid(),
-                        },
-                        usersLength: 1,
-                        tracks: [
-                            {
-                                id: datatype.uuid(),
-                                artistName: name.findName(),
-                                duration: 42000,
-                                title: random.words(3),
-                                score: datatype.number(),
-                            },
-                        ],
-                        suggestedTracks: null,
-                    },
-                };
-            });
-        sinon.stub(ServerToTemporalController, 'play').callsFake(async () => {
-            userCouldEmitAnExclusiveRoomSignal = true;
-            console.log('play mock called');
-            return;
-        });
-        sinon
-            .stub(ServerToTemporalController, 'terminateWorkflow')
-            .callsFake(async (): Promise<void> => {
-                return;
-            });
-        /** ***** */
-
-        /**
-         * User connects two devices then create a room from one
-         */
-        const socketA = await createUserAndGetSocket({ userID });
-        const socketB = await createSocketConnection({ userID });
-        assert.equal((await Device.all()).length, 2);
-        socketA.emit('CREATE_ROOM', {
-            name: roomName,
-            initialTracksIDs: [datatype.uuid()],
-        });
-        await sleep();
-        assert.isNotNull(await MtvRoom.findBy('creator', userID));
-
-        /**
-         * From the other one he emits an ACTION_PLAY event
-         * It achieves only if he joined the socket io server room
-         */
-        socketB.emit('ACTION_PLAY');
-        await sleep();
-        assert.equal(userCouldEmitAnExclusiveRoomSignal, true);
-    });
-
-    test('New user socket connection should join previously joined/created room', async (assert) => {
-        const creatorUserID = datatype.uuid();
-        const roomName = random.word();
-        const socketA = await createUserAndGetSocket({
-            userID: creatorUserID,
-        });
-        let userCouldEmitAnExclusiveRoomSignal = false;
-        /** Mocks */
-        sinon
-            .stub(ServerToTemporalController, 'createMtvWorkflow')
-            .callsFake(async ({ workflowID }) => {
-                const creator = datatype.uuid();
-                const state: MtvWorkflowStateWithUserRelatedInformation = {
-                    roomID: workflowID,
-                    roomCreatorUserID: creatorUserID,
-                    playing: false,
-                    name: roomName,
-                    usersLength: 1,
-                    currentTrack: null,
-                    userRelatedInformation: {
-                        userID: creator,
-                        emittingDeviceID: datatype.uuid(),
-                    },
-                    tracks: [
-                        {
-                            id: datatype.uuid(),
-                            artistName: name.findName(),
-                            duration: 42000,
-                            title: random.words(3),
-                            score: datatype.number(),
-                        },
-                    ],
-                    suggestedTracks: null,
-                };
-
-                return {
-                    runID: datatype.uuid(),
-                    workflowID,
-                    state,
-                };
-            });
-        sinon
-            .stub(ServerToTemporalController, 'getState')
-            .callsFake(async ({ workflowID, userID }) => {
-                return {
-                    name: roomName,
-                    roomCreatorUserID: creatorUserID,
-                    playing: false,
-                    roomID: workflowID,
-                    usersLength: 1,
-                    userRelatedInformation: userID
-                        ? {
-                              userID,
-                              emittingDeviceID: datatype.uuid(),
-                          }
-                        : null,
-                    currentTrack: null,
-                    tracks: null,
-                    suggestedTracks: null,
-                };
-            });
-        sinon.stub(ServerToTemporalController, 'play').callsFake(async () => {
-            userCouldEmitAnExclusiveRoomSignal = true;
-            return;
-        });
-        /** ***** */
-
-        /**
-         * User connects one device, then creates a room from it
-         */
-        socketA.emit('CREATE_ROOM', {
-            name: roomName,
-            initialTracksIDs: [datatype.uuid()],
-        });
-        await sleep();
-
-        /**
-         * User connects a new device, then emits an ACTION_PLAY
-         * It achieves only if he joined the socket io server room
-         * He also receives the mtvRoom's context
-         */
-        const socketB = {
-            socket: await createSocketConnection({ userID: creatorUserID }),
-        };
-
-        socketB.socket.emit('ACTION_PLAY');
-        await sleep();
-
-        assert.isTrue(userCouldEmitAnExclusiveRoomSignal);
-    });
-
-    test('Go to next tracks events are forwarded to Temporal', async (assert) => {
-        /**
-         * Create a user that is member of a mtv room.
-         * We want this user to send a GO_TO_NEXT_TRACK event and assert
-         * that the method that forwards the event is correctly called.
-         */
-        const userID = datatype.uuid();
-
-        const mtvRoomIDToAssociate = datatype.uuid();
-        const socket = await createUserAndGetSocket({
-            userID,
-            mtvRoomIDToAssociate,
-        });
-
-        const goToNextTrackStub = sinon
-            .stub(ServerToTemporalController, 'goToNextTrack')
-            .resolves();
-
-        socket.emit('GO_TO_NEXT_TRACK');
-
-        await sleep();
-
-        assert.isTrue(goToNextTrackStub.calledOnce);
-    });
-
-    test('It should send back the socket related mtv room context', async (assert) => {
-        /**
-         * Manually create and associate room to user and socket to user
-         * Then emit a GET_CONTEXT and verify that the test achieves to it
-         */
-        const userID = datatype.uuid();
-        const mtvRoomIDToAssociate = datatype.uuid();
-        const socket = await createUserAndGetSocket({
-            userID,
-            mtvRoomIDToAssociate,
-        });
-        assert.equal(1, (await MtvRoom.all()).length);
-        const receivedEvents: string[] = [];
-
-        socket.once('RETRIEVE_CONTEXT', () => {
-            receivedEvents.push('RETRIEVE_CONTEXT');
-        });
-
-        sinon
-            .stub(ServerToTemporalController, 'getState')
-            .callsFake(async () => {
-                return {
-                    roomID: mtvRoomIDToAssociate,
-                    currentTrack: null,
-                    name: random.word(),
-                    playing: false,
-
-                    roomCreatorUserID: userID,
-                    tracks: null,
-                    suggestedTracks: null,
-                    userRelatedInformation: null,
-                    usersLength: 1,
-                };
-            });
-
-        socket.emit('GET_CONTEXT');
-
-        await sleep();
-
-        assert.equal(receivedEvents[0], 'RETRIEVE_CONTEXT');
-    });
-
-    test('It throw an error as no user device could join socket room', async (assert) => {
-        const userID = datatype.uuid();
-        const user = await User.create({
-            uuid: userID,
-            nickname: random.word(),
-        });
-        const roomID = datatype.uuid();
-
-        const devices = await Device.createMany([
-            {
-                userID: userID,
-                uuid: datatype.uuid(),
-                socketID: datatype.uuid(),
-                name: random.word(),
-            },
-            {
-                userID: userID,
-                uuid: datatype.uuid(),
-                socketID: datatype.uuid(),
-                name: random.word(),
-            },
-        ]);
-
-        user.related('devices').saveMany(devices);
-
-        try {
-            await UserService.joinEveryUserDevicesToRoom(user, roomID);
-        } catch ({ message }) {
-            assert.equal(
-                message,
-                `couldn't join for any device for user ${user.uuid}`,
-            );
-        }
-    }).timeout(10_000);
-
-    test('It should send server socket event to the client via the UserService.emitEventInSocket', async (assert) => {
-        const userID = datatype.uuid();
-        const socket = await createUserAndGetSocket({ userID });
-        const state: MtvWorkflowState = {
-            currentTrack: null,
-            name: random.word(),
-            playing: false,
-            roomCreatorUserID: userID,
-            roomID: datatype.uuid(),
-            tracks: null,
-            suggestedTracks: null,
-            usersLength: 1,
-            userRelatedInformation: null,
-        };
-        const receivedEvents: string[] = [];
-
-        socket.once('CREATE_ROOM_CALLBACK', (payload) => {
-            assert.deepEqual(payload, state);
-            receivedEvents.push('CREATE_ROOM_CALLBACK');
-        });
-
-        UserService.emitEventInSocket(socket.id, 'CREATE_ROOM_CALLBACK', [
-            state,
-        ]);
-
-        await sleep();
-        assert.isTrue(receivedEvents.includes('CREATE_ROOM_CALLBACK'));
     });
 });
